Planing Searching Heuristic Analysis
This project includes skeletons for the classes and functions needed to
solve deterministic logistics planning problems for an Air Cargo
transport system using a planning search agent. And with the increase of
state and action variables , the complexity of the problem is
exponential. In this analysis report , we first make an analysis to the
complexity for the problem . Second we compare with different
heuristics functions and find which is better.